Refactor component registration to use centralized linking
Moved component registration from individual component files to a new ComponentForcelink class and macro (COMP_REG), which is now invoked centrally. This change simplifies and centralizes the registration process, making it easier to manage and extend. Also updated EditorContext to write engine path to APPDATA, and made minor improvements to TextRenderer and includes.